requirements.txt
line-bot-sdk==3.7.0
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0